+++ src/main/java/org/realityforge/replicant/client/transport/ClientSession.java
@@ -1,15 +1,30 @@
 package org.realityforge.replicant.client.transport;
 
+import java.util.LinkedList;
 import javax.annotation.Nonnull;
 
 /**
  * Abstract representation of client session.
- * Simply tracks the session identifier and job sequencing.
  */
 public abstract class ClientSession
 {
   private final String _sessionID;
   private final RequestManager _requestManager;
+
+  /**
+   * The set of data load actions that still need to have the json parsed.
+   */
+  private final LinkedList<DataLoadAction> _pendingActions = new LinkedList<>();
+  /**
+   * The set of data load actions that have their json parsed. They are inserted into
+   * this list according to their sequence.
+   */
+  private final LinkedList<DataLoadAction> _parsedActions = new LinkedList<>();
+  /**
+   * Sometimes a data load action occurs that is not initiated by the server. These do not
+   * typically need to be sequenced and are prioritized above other actions.
+   */
+  private final LinkedList<DataLoadAction> _oobActions = new LinkedList<>();
 
   private int _lastRxSequence;
 
@@ -29,6 +44,21 @@ public abstract class ClientSession
   public RequestManager getRequestManager()
   {
     return _requestManager;
+  }
+
+  final LinkedList<DataLoadAction> getPendingActions()
+  {
+    return _pendingActions;
+  }
+
+  final LinkedList<DataLoadAction> getParsedActions()
+  {
+    return _parsedActions;
+  }
+
+  final LinkedList<DataLoadAction> getOobActions()
+  {
+    return _oobActions;
   }
 
   protected RequestManager newRequestManager()